check 7

+++ slither/detectors/my_detectors/tcheck.py
@@ -2373,7 +2373,7 @@ def convert_ssa_name(name):
                 pos = revpos
                 break
     _name = name[:pos]
-    num = (int)(name[pos+1:])
+    num = (name[pos+1:])
     print(f"Name: {_name}, Num: {num}")
     return [_name, num]
